Minor corrections on tests
